add water physic and swimming action
