+++ index.js
@@ -5,10 +5,11 @@ const {useApp, useFrame, useCamera, useLocalPlayer, usePhysics, useProcGenManage
 const {GPUTaskManager} = useGPUTask();
 const {GenerationTaskManager} = useGenerationTask();
 
-import {TerrainMesh} from './terrain-mesh.js';
-import {WaterMesh} from './water-mesh.js';
-import {BarrierMesh} from './barrier-mesh.js';
-import {LitterMetaMesh} from './litter-mesh.js';
+import {TerrainMesh} from './layers/terrain-mesh.js';
+import {WaterMesh} from './layers/water-mesh.js';
+// import {BarrierMesh} from './layers/barrier-mesh.js';
+// import {LitterMetaMesh, litterUrls} from './layers/litter-mesh.js';
+// import {GrassMesh, grassUrls} from './layers/grass-mesh.js';
 
 // locals
 
@@ -18,45 +19,6 @@ const localVector3 = new THREE.Vector3();
 const localQuaternion = new THREE.Quaternion();
 const localMatrix = new THREE.Matrix4();
 const localMatrix2 = new THREE.Matrix4();
-
-// urls
-
-const procgenAssetsBaseUrl = `https://webaverse.github.io/procgen-assets/`;
-const urlSpecs = {
-  trees: [
-    `Tree_1_1.glb`,
-    `Tree_1_2.glb`,
-    `Tree_2_1.glb`,
-    `Tree_2_2.glb`,
-    `Tree_3_1.glb`,
-    `Tree_3_2.glb`,
-    `Tree_4_1.glb`,
-    `Tree_4_2.glb`,
-    `Tree_4_3.glb`,
-    `Tree_5_1.glb`,
-    `Tree_5_2.glb`,
-    `Tree_6_1.glb`,
-    `Tree_6_2.glb`,
-  ].map(u => {
-    return `${procgenAssetsBaseUrl}vegetation/garden-trees/${u}`;
-  }),
-  ores: [
-    `BlueOre_deposit_low.glb`,
-    `Iron_Deposit_low.glb`,
-    `Ore_Blue_low.glb`,
-    `Ore_BrownRock_low.glb`,
-    `Ore_Deposit_Red.glb`,
-    `Ore_Red_low.glb`,
-    `Ore_metal_low.glb`,
-    `Ore_wood_low.glb`,
-    `Rock_ore_Deposit_low.glb`,
-    `TreeOre_low.glb`,
-  ].map(u => {
-    return `${procgenAssetsBaseUrl}/litter/ores/${u}`;
-  }),
-};
-const litterUrls = urlSpecs.trees.slice(0, 1)
-  .concat(urlSpecs.ores.slice(0, 1));
 
 // main
 
@@ -86,7 +48,7 @@ export default e => {
     // lodTracker.debugMesh.updateMatrixWorld();
 
     lodTracker.onPostUpdate(currentCoord => {
-      barrierMesh.updateChunk(currentCoord);
+      // barrierMesh.updateChunk(currentCoord);
     });
 
     // managers
@@ -112,21 +74,29 @@ export default e => {
     app.add(waterMesh);
     waterMesh.updateMatrixWorld();
 
-    const barrierMesh = new BarrierMesh({
-      instance,
-      gpuTaskManager,
-    });
-    barrierMesh.frustumCulled = false;
-    app.add(barrierMesh);
-    barrierMesh.updateMatrixWorld();
+    // const barrierMesh = new BarrierMesh({
+    //   instance,
+    //   gpuTaskManager,
+    // });
+    // barrierMesh.frustumCulled = false;
+    // app.add(barrierMesh);
+    // barrierMesh.updateMatrixWorld();
 
-    const litterMesh = new LitterMetaMesh({
-      instance,
-      gpuTaskManager,
-      physics,
-    });
-    app.add(litterMesh);
-    litterMesh.updateMatrixWorld();
+    // const litterMesh = new LitterMetaMesh({
+    //   instance,
+    //   gpuTaskManager,
+    //   physics,
+    // });
+    // app.add(litterMesh);
+    // litterMesh.updateMatrixWorld();
+
+    // const grassMesh = new GrassMesh({
+    //   instance,
+    //   gpuTaskManager,
+    //   physics,
+    // });
+    // app.add(grassMesh);
+    // grassMesh.updateMatrixWorld();
 
     // genration events handling
     lodTracker.onChunkAdd(async chunk => {
@@ -135,42 +105,57 @@ export default e => {
       const generation = generationTaskManager.createGeneration(key);
       generation.addEventListener('geometryadd', e => {
         const {result} = e.data;
-        const {heightfield, vegetation} = result;
+        const {heightfield, vegetation, grass} = result;
         
         // heightfield
         terrainMesh.addChunk(chunk, heightfield);
         waterMesh.addChunk(chunk, heightfield);
-        barrierMesh.addChunk(chunk, heightfield);
+        // barrierMesh.addChunk(chunk, heightfield);
       
         // vegetation
-        litterMesh.addChunk(chunk, vegetation);
+        // litterMesh.addChunk(chunk, vegetation);
+        
+        // grass
+        // grassMesh.addChunk(chunk, grass);
       });
       generation.addEventListener('geometryremove', e => {
         // heightfield
         terrainMesh.removeChunk(chunk);
         waterMesh.removeChunk(chunk);
-        barrierMesh.removeChunk(chunk);
+        // barrierMesh.removeChunk(chunk);
 
         // vegetation
-        litterMesh.removeChunk(chunk);
+        // litterMesh.removeChunk(chunk);
+
+        // grass
+        // grassMesh.removeChunk(chunk);
       });
 
       try {
         const signal = generation.getSignal();
+        const generateFlags = {
+          terrain: true,
+          water: true,
+          // barrier: true,
+          // vegetation: true,
+          // grass: true,
+        };
+        const options = {
+          signal,
+        };
         const [
           heightfield,
           vegetation,
+          grass,
         ] = await Promise.all([
-          instance.generateChunk(chunk.min, chunk.lod, chunk.lodArray, {
-            signal,
-          }),
-          instance.generateVegetation(chunk.min, chunk.lod, litterUrls.length, {
-            signal,
-          }),
+          instance.generateChunk(chunk.min, chunk.lod, chunk.lodArray, generateFlags, options),
+          // instance.generateVegetation(chunk.min, chunk.lod, litterUrls.length, options),
+          // instance.generateGrass(chunk.min, chunk.lod, grassUrls.length, options),
         ]);
         generation.finish({
           heightfield,
-          vegetation,
+          // vegetation,
+          // grass,
         });
       } catch (err) {
         if (err.isAbortError) {
@@ -189,11 +174,16 @@ export default e => {
     });
 
     // load
-    await litterMesh.loadUrls(litterUrls);
+    const _waitForLoad = async () => {
+      await Promise.all([
+        // litterMesh.waitForLoad(),
+        // grassMesh.waitForLoad(),
+      ]);
+    };
+    await _waitForLoad();
 
     // frame handling
     frameCb = () => {
-      
       const _updateLodTracker = () => {
         const localPlayer = useLocalPlayer();
 
@@ -223,7 +213,7 @@ export default e => {
       _updateLodTracker();
 
       const _updateLitteMesh = () => {
-        litterMesh.update();
+        // litterMesh.update(); // update spritesheet uniforms
       };
       _updateLitteMesh();
