+++ layers/water-mesh.js
@@ -2,22 +2,26 @@ import * as THREE from 'three';
 import metaversefile from 'metaversefile';
 import {bufferSize, WORLD_BASE_HEIGHT, MIN_WORLD_HEIGHT, MAX_WORLD_HEIGHT} from '../constants.js';
 
-const {useProcGenManager, useGeometryBuffering} = metaversefile;
+const {useProcGenManager, useGeometryBuffering, useLocalPlayer} = metaversefile;
 const {BufferedMesh, GeometryAllocator} = useGeometryBuffering();
 const procGenManager = useProcGenManager();
 
 //
-
+const fakeMaterial = new THREE.MeshBasicMaterial({
+  color: 0xffffff,
+});
 const localVector3D = new THREE.Vector3();
 const localVector3D2 = new THREE.Vector3();
 const localBox = new THREE.Box3();
-
+const localQuaternion = new THREE.Quaternion();
+const localVector = new THREE.Vector3();
 //
 
 export class WaterMesh extends BufferedMesh {
   constructor({
     instance,
     gpuTaskManager,
+    physics
   }) {
     const allocator = new GeometryAllocator(
       [
@@ -47,7 +51,7 @@ export class WaterMesh extends BufferedMesh {
     const {geometry} = allocator;
     const material = new THREE.MeshNormalMaterial();
 
-    super(geometry, material);
+    super(geometry);
 
     this.instance = instance;
     this.gpuTaskManager = gpuTaskManager;
@@ -55,6 +59,14 @@ export class WaterMesh extends BufferedMesh {
     this.allocator = allocator;
     this.gpuTasks = new Map();
     this.geometryBindings = new Map();
+
+    this.material = new THREE.MeshBasicMaterial( {color: 0x0000ff, side: THREE.DoubleSide, transparent: true, opacity: 0.9} );
+    this.geometry = geometry;
+    this.physics = physics;
+    this.physicsObjectsMap = new Map();
+    this.currentChunkMap = new Map();
+    this.currentWaterHeightMap = new Map();
+    this.lastUpdateCoord = new THREE.Vector2();
   }
   addChunk(chunk, chunkResult) {
     const key = procGenManager.getNodeHash(chunk);
@@ -156,33 +168,41 @@ export class WaterMesh extends BufferedMesh {
 
         this.geometryBindings.set(key, geometryBinding);
       };
-      _handleWaterMesh(chunkResult.waterGeometry);
+      const waterGeometry = chunkResult.waterGeometry
+      _handleWaterMesh(waterGeometry);
 
-      /* const _handlePhysics = async () => {
-        if (geometryBuffer) {
-          this.matrixWorld.decompose(localVector, localQuaternion, localVector2);
+      const _handlePhysics = async () => {
+        const physicsGeo = new THREE.BufferGeometry();
+        physicsGeo.setAttribute(
+          'position',
+          new THREE.BufferAttribute(waterGeometry.positions, 3)
+        );
+        physicsGeo.setIndex(
+          new THREE.BufferAttribute(waterGeometry.indices, 1)
+        );
+        const physicsMesh = new THREE.Mesh(physicsGeo, fakeMaterial);
+        
+        const geometryBuffer = await this.physics.cookGeometryAsync(physicsMesh);
+
+        if (geometryBuffer && geometryBuffer.length !== 0) {
+          this.matrixWorld.decompose(
+            localVector3D,
+            localQuaternion,
+            localVector3D2
+          );
           const physicsObject = this.physics.addCookedGeometry(
             geometryBuffer,
-            localVector,
+            localVector3D,
             localQuaternion,
-            localVector2
+            localVector3D2
           );
-          this.physicsObjects.push(physicsObject);
-          this.physicsObjectToChunkMap.set(physicsObject, chunk);
-
-          const onchunkremove = () => {
-            this.physics.removeGeometry(physicsObject);
-
-            const index = this.physicsObjects.indexOf(physicsObject);
-            this.physicsObjects.splice(index, 1);
-            this.physicsObjectToChunkMap.delete(physicsObject);
-
-            tracker.offChunkRemove(chunk, onchunkremove);
-          };
-          tracker.onChunkRemove(chunk, onchunkremove);
+          this.physics.disableGeometryQueries(physicsObject); // disable each physicsObject
+          this.physicsObjectsMap.set(key, physicsObject);
+          this.currentChunkMap.set(chunk.min.x + ',' + chunk.min.y, physicsObject); // use string of chunk.min as a key to map each physicsObject
+          this.currentWaterHeightMap.set(chunk.min.x + ',' + chunk.min.y, waterGeometry.positions[1]); // use string of chunk.min as a key to map the posY of each chunk
         }
       };
-      _handlePhysics(); */
+      _handlePhysics();
     });
     this.gpuTasks.set(key, task);
   }
@@ -204,6 +224,72 @@ export class WaterMesh extends BufferedMesh {
       const task = this.gpuTasks.get(key);
       task.cancel();
       this.gpuTasks.delete(key);
+    }
+  }
+  update() {
+    const localPlayer = useLocalPlayer();
+    const lastUpdateCoordKey = this.lastUpdateCoord.x + ',' + this.lastUpdateCoord.y; 
+    const currentChunkPhysicObject = this.currentChunkMap.get(lastUpdateCoordKey); // use lodTracker.lastUpdateCoord as a key to check which chunk player currently at 
+    const waterSurfacePos = this.currentWaterHeightMap.get(lastUpdateCoordKey); // use lodTracker.lastUpdateCoord as a key to check the pos y of the current chunk
+    let contactWater = false;
+    if (currentChunkPhysicObject) { // if we get the physicObject of the current chunk, then use overlapBox to check whether player contact the water
+      this.physics.enableGeometryQueries(currentChunkPhysicObject);
+      if (localPlayer.avatar) {
+        let collisionIds;
+        const height = localPlayer.avatar.height * 0.9;
+        const width = localPlayer.avatar.shoulderWidth
+        if (localPlayer.position.y > waterSurfacePos) {
+          collisionIds = this.physics.overlapBox(width, height, width, localPlayer.position, localPlayer.quaternion).objectIds;
+        }
+        else {
+          localVector.set(localPlayer.position.x, waterSurfacePos, localPlayer.position.z);
+          collisionIds = this.physics.overlapBox(width, height, width, localVector, localPlayer.quaternion).objectIds;
+        } 
+        for (const collisionId of collisionIds) {
+          if (collisionId === currentChunkPhysicObject.physicsId)
+            contactWater = true;
+        }
+      }
+      this.physics.disableGeometryQueries(currentChunkPhysicObject);
+    }
+
+    // handle swimming action
+    if (contactWater) {
+      this.material.color.setHex( 0x0000ff ); // for testing
+      if(waterSurfacePos >= localPlayer.position.y - localPlayer.avatar.height + localPlayer.avatar.height * 0.8){
+        if(!localPlayer.hasAction('swim')){
+          //console.log('add');
+          const swimAction = {
+              type: 'swim',
+              onSurface: false,
+              swimDamping: 1,
+              animationType: 'breaststroke'
+          };
+          localPlayer.setControlAction(swimAction);
+        }
+
+        if (waterSurfacePos < localPlayer.position.y - localPlayer.avatar.height + localPlayer.avatar.height * 0.85) {
+          if (localPlayer.hasAction('swim') && !localPlayer.getAction('swim').onSurface) {
+            localPlayer.getAction('swim').onSurface = true;
+          }
+        }
+        else {
+          if (localPlayer.hasAction('swim') && localPlayer.getAction('swim').onSurface) {
+            localPlayer.getAction('swim').onSurface = false;
+          }
+        }
+      }
+      else{
+          if (localPlayer.hasAction('swim')) {
+            localPlayer.removeAction('swim');
+          }
+      }  
+    } 
+    else {
+      this.material.color.setHex( 0xff0000 ); // for testing
+      if (localPlayer.hasAction('swim')) {
+        localPlayer.removeAction('swim');
+      }
     }
   }
 }
